# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "The Open Service Broker API defines an HTTP(S) interface between Platforms\
    \ and Service Brokers."
  version: "master - might contain changes that are not yet released"
  title: "Open Service Broker API"
  contact:
    name: "Open Service Broker API"
    url: "https://www.openservicebrokerapi.org/"
    email: "open-service-broker-api@googlegroups.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
- basicAuth: []
paths:
...
  /v2/service_instances/{instance_id}:
    get:
      tags:
      - "ServiceInstances"
      summary: "gets a service instance"
      operationId: "serviceInstance.get"
      parameters:
      - name: "X-Broker-API-Version"
        in: "header"
        description: "version number of the Service Broker API that the Platform will\
          \ use"
        required: true
        type: "string"
      - name: "X-Broker-API-Originating-Identity"
        in: "header"
        description: "identity of the user that initiated the request from the Platform"
        required: false
        type: "string"
      - name: "X-Broker-API-Request-Identity"
        in: "header"
        description: "idenity of the request from the Platform"
        required: false
        type: "string"
      - name: "instance_id"
        in: "path"
        description: "instance id of instance to provision"
        required: true
        type: "string"
      - name: "service_id"
        in: "query"
        description: "id of the service associated with the instance"
        required: false
        type: "string"
      - name: "plan_id"
        in: "query"
        description: "id of the plan associated with the instance"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/ServiceInstanceResource"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        "412":
          description: "Precondition Failed"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceInstances"
    put:
      tags:
      - "ServiceInstances"
      summary: "provision a service instance"
      operationId: "serviceInstance.provision"
      parameters:
      - name: "X-Broker-API-Version"
        in: "header"
        description: "version number of the Service Broker API that the Platform will\
          \ use"
        required: true
        type: "string"
      - name: "X-Broker-API-Originating-Identity"
        in: "header"
        description: "identity of the user that initiated the request from the Platform"
        required: false
        type: "string"
      - name: "X-Broker-API-Request-Identity"
        in: "header"
        description: "idenity of the request from the Platform"
        required: false
        type: "string"
      - name: "instance_id"
        in: "path"
        description: "instance id of instance to provision"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "parameters for the requested service instance provision"
        required: true
        schema:
          $ref: "#/definitions/ServiceInstanceProvisionRequest"
      - name: "accepts_incomplete"
        in: "query"
        description: "asynchronous operations supported"
        required: false
        type: "boolean"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/ServiceInstanceProvisionResponse"
        "201":
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceInstanceProvisionResponse"
        "202":
          description: "Accepted"
          schema:
            $ref: "#/definitions/ServiceInstanceAsyncOperation"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        "409":
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        "412":
          description: "Precondition Failed"
          schema:
            $ref: "#/definitions/Error"
        "422":
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceInstances"
    delete:
      tags:
      - "ServiceInstances"
      summary: "deprovision a service instance"
      operationId: "serviceInstance.deprovision"
      parameters:
      - name: "X-Broker-API-Version"
        in: "header"
        description: "version number of the Service Broker API that the Platform will\
          \ use"
        required: true
        type: "string"
      - name: "X-Broker-API-Originating-Identity"
        in: "header"
        description: "identity of the user that initiated the request from the Platform"
        required: false
        type: "string"
      - name: "X-Broker-API-Request-Identity"
        in: "header"
        description: "idenity of the request from the Platform"
        required: false
        type: "string"
      - name: "instance_id"
        in: "path"
        description: "instance id of instance to provision"
        required: true
        type: "string"
      - name: "accepts_incomplete"
        in: "query"
        description: "asynchronous operations supported"
        required: false
        type: "boolean"
      - name: "service_id"
        in: "query"
        description: "id of the service associated with the instance being deleted"
        required: true
        type: "string"
      - name: "plan_id"
        in: "query"
        description: "id of the plan associated with the instance being deleted"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
        "202":
          description: "Accepted"
          schema:
            $ref: "#/definitions/AsyncOperation"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        "410":
          description: "Gone"
          schema:
            $ref: "#/definitions/Error"
        "412":
          description: "Precondition Failed"
          schema:
            $ref: "#/definitions/Error"
        "422":
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceInstances"
    patch:
      tags:
      - "ServiceInstances"
      summary: "update a service instance"
      operationId: "serviceInstance.update"
      parameters:
      - name: "X-Broker-API-Version"
        in: "header"
        description: "version number of the Service Broker API that the Platform will\
          \ use"
        required: true
        type: "string"
      - name: "X-Broker-API-Originating-Identity"
        in: "header"
        description: "identity of the user that initiated the request from the Platform"
        required: false
        type: "string"
      - name: "X-Broker-API-Request-Identity"
        in: "header"
        description: "idenity of the request from the Platform"
        required: false
        type: "string"
      - name: "instance_id"
        in: "path"
        description: "instance id of instance to provision"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "parameters for the requested service instance update"
        required: true
        schema:
          $ref: "#/definitions/ServiceInstanceUpdateRequest"
      - name: "accepts_incomplete"
        in: "query"
        description: "asynchronous operations supported"
        required: false
        type: "boolean"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
        "202":
          description: "Accepted"
          schema:
            $ref: "#/definitions/ServiceInstanceAsyncOperation"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        "412":
          description: "Precondition Failed"
          schema:
            $ref: "#/definitions/Error"
        "422":
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceInstances"
...
securityDefinitions:
  basicAuth:
    type: "basic"
definitions:
...
  ServiceInstanceResource:
    type: "object"
    properties:
      service_id:
        type: "string"
      plan_id:
        type: "string"
      dashboard_url:
        type: "string"
      parameters:
        type: "object"
        properties: {}
      maintenance_info:
        $ref: "#/definitions/MaintenanceInfo"
      metadata:
        $ref: "#/definitions/ServiceInstanceMetadata"
    example:
      metadata:
        attributes: "{}"
        labels: "{}"
      maintenance_info:
        description: "description"
        version: "version"
      service_id: "service_id"
      dashboard_url: "dashboard_url"
      parameters: "{}"
      plan_id: "plan_id"
  ServiceInstanceProvisionRequest:
    type: "object"
    required:
    - "organization_guid"
    - "plan_id"
    - "service_id"
    - "space_guid"
    properties:
      service_id:
        type: "string"
      plan_id:
        type: "string"
      context:
        $ref: "#/definitions/Context"
      organization_guid:
        type: "string"
        x-deprecated: true
      space_guid:
        type: "string"
        x-deprecated: true
      parameters:
        type: "object"
        properties: {}
      maintenance_info:
        $ref: "#/definitions/MaintenanceInfo"
    example:
      maintenance_info:
        description: "description"
        version: "version"
      service_id: "service_id"
      context: {}
      organization_guid: "organization_guid"
      space_guid: "space_guid"
      parameters: "{}"
      plan_id: "plan_id"
  ServiceInstanceProvisionResponse:
    type: "object"
    properties:
      dashboard_url:
        type: "string"
      metadata:
        $ref: "#/definitions/ServiceInstanceMetadata"
    example:
      metadata:
        attributes: "{}"
        labels: "{}"
      dashboard_url: "dashboard_url"
  ServiceInstanceAsyncOperation:
    type: "object"
    properties:
      dashboard_url:
        type: "string"
      operation:
        type: "string"
      metadata:
        $ref: "#/definitions/ServiceInstanceMetadata"
  ServiceInstanceUpdateRequest:
    type: "object"
    required:
    - "service_id"
    properties:
      context:
        $ref: "#/definitions/Context"
      service_id:
        type: "string"
      plan_id:
        type: "string"
      parameters:
        type: "object"
        properties: {}
      previous_values:
        $ref: "#/definitions/ServiceInstancePreviousValues"
      maintenance_info:
        $ref: "#/definitions/MaintenanceInfo"
    example:
      maintenance_info:
        description: "description"
        version: "version"
      previous_values:
        maintenance_info:
          description: "description"
          version: "version"
        service_id: "service_id"
        organization_id: "organization_id"
        space_id: "space_id"
        plan_id: "plan_id"
      service_id: "service_id"
      context: {}
      parameters: "{}"
      plan_id: "plan_id"
  ServiceInstancePreviousValues:
    type: "object"
    properties:
      service_id:
        type: "string"
        x-deprecated: true
      plan_id:
        type: "string"
      organization_id:
        type: "string"
        x-deprecated: true
      space_id:
        type: "string"
        x-deprecated: true
      maintenance_info:
        $ref: "#/definitions/MaintenanceInfo"
    example:
      maintenance_info:
        description: "description"
        version: "version"
      service_id: "service_id"
      organization_id: "organization_id"
      space_id: "space_id"
      plan_id: "plan_id"
  AsyncOperation:
    type: "object"
    properties:
      operation:
        type: "string"
...
  ServiceInstanceMetadata:
    type: "object"
    properties:
      labels:
        type: "object"
        properties: {}
      attributes:
        type: "object"
        properties: {}
    example:
      attributes: "{}"
      labels: "{}"
  MaintenanceInfo:
    type: "object"
    properties:
      version:
        type: "string"
      description:
        type: "string"
    example:
      description: "description"
      version: "version"
  Context:
    type: "object"
    description: "See [Context Conventions](https://github.com/openservicebrokerapi/servicebroker/blob/master/profile.md#context-object)\
      \ for more details."
...
  Error:
    type: "object"
    properties:
      error:
        type: "string"
      description:
        type: "string"
      instance_usable:
        type: "boolean"
      update_repeatable:
        type: "boolean"
...